# apps/daemon/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "notemap",
    "version": "1",
    "description": "The capture-and-feed subset of notemap's /v1 surface. Binds to localhost by default; no authentication."
  },
  "components": {
    "schemas": {
...
      "Session": {
        "type": "object",
        "properties": {
          "authenticated": {
            "type": "boolean",
            "example": true
          },
          "requiresCredentials": {
            "type": "boolean",
            "example": true
          },
          "identity": {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "session",
                  "token"
                ],
                "example": "session"
              },
              "id": {
                "type": "string",
                "example": "gpeukvybsmmgwnec"
              },
              "name": {
                "type": "string",
                "example": "laptop"
              }
            },
            "required": [
              "kind",
              "id"
            ]
          }
        },
        "required": [
          "authenticated",
          "requiresCredentials"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "example": "anton"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "example": "correct horse battery staple"
          }
        },
        "required": [
          "name",
          "password"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/session": {
      "post": {
        "summary": "Sign in and take a session",
        "description": "Exchanges the credential for a session cookie. The cookie is `HttpOnly` and carries the only copy of the session's secret; the daemon stores a hash of it and can never reproduce it. Answered the same way whether the name or the password was wrong, so neither can be probed for.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in. The session cookie is set.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "The body could not be read.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "malformed-json",
                            "malformed-envelope"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ],
                      "additionalProperties": {},
                      "description": "The refusal's kind, with its facts beside it."
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The credential was not accepted. Which half was wrong is not said.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "unauthenticated"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ],
                      "additionalProperties": {},
                      "description": "The refusal's kind, with its facts beside it."
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "415": {
            "description": "The body was not JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "unsupported-media-type"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ],
                      "additionalProperties": {},
                      "description": "The refusal's kind, with its facts beside it."
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Too many attempts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "too-many-attempts"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ],
                      "additionalProperties": {},
                      "description": "The refusal's kind, with its facts beside it."
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Read who this request is, if anyone",
        "description": "Open, and answers 200 whether or not anyone is signed in — being signed out is an answer rather than a refusal, and a client needs to tell it apart from a daemon that is unreachable. `requiresCredentials` is false on a daemon nobody has set a password on, where every request is let through.",
        "responses": {
          "200": {
            "description": "What this request is, and whether this daemon asks at all.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Sign out",
        "description": "Ends the session the cookie names and clears the cookie. Idempotent, and open: signing out with a session that already lapsed is not an error, it is the same outcome arrived at early.",
        "responses": {
          "204": {
            "description": "Signed out, whether or not there was a session."
          },
          "422": {
            "description": "Authenticated by an access token, which is not a session to end.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "not-a-session"
                          ]
                        }
                      },
                      "required": [
                        "code"
                      ],
                      "additionalProperties": {},
                      "description": "The refusal's kind, with its facts beside it."
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}